+++ CHANGELOG.md
@@ -4,15 +4,26 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Changelog](http://keepachangelog.com/).
 
 
+
 ## Unreleased
 ---
 
 ### New
-* DataSource: `settings` [#4](https://github.com/brittandeyoung/terraform-provider-awsteam/issues/4)
-* Resource: `settings` [#4](https://github.com/brittandeyoung/terraform-provider-awsteam/issues/4)
 
 ### Changes
 
 ### Fixes
 
 ### Breaks
+
+
+## 0.1.0 - (2024-01-12)
+---
+
+### New
+* Resource: `awsteam_approvers_account` [#7](https://github.com/brittandeyoung/terraform-provider-awsteam/issues/7)
+* Resource: `awsteam_approvers_ou` [#7](https://github.com/brittandeyoung/terraform-provider-awsteam/issues/7)
+* DataSource: `awsteam_settings` [#4](https://github.com/brittandeyoung/terraform-provider-awsteam/issues/4)
+* Resource: `awsteam_settings` [#4](https://github.com/brittandeyoung/terraform-provider-awsteam/issues/4)
+
+